Partially revert #763 for consistency
memory_storage.* now use the same pattern as other storages
